authentication guards and authentication service fixes

+++ frontend/Umbral.WebPortal/src/app/core/services/local-storage/local-storage.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from "@angular/core";
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+import { isPlatformBrowser } from "@angular/common";
+import { inject, Injectable, PLATFORM_ID } from "@angular/core";
+import { Utilities } from "../../utilities/utilities";
 
 /**
  * Local storage service
@@ -6,16 +10,29 @@ import { Injectable } from "@angular/core";
 @Injectable({ providedIn: 'root' })
 export class LocalStorageService {
 
+    private readonly _platformId: any = inject(PLATFORM_ID);
+
+    /**
+     * Whether we are executing on the browser.
+     */
+    private get isBrowser(): boolean {
+        return isPlatformBrowser(this._platformId);
+    }
+
     /**
      * Sets an item by a given key in the local storage.
      * @param key 
      * @param value 
      */
-    public setItem(key: string, value: any): void {
+    public setItem<TObject>(key: string, value: TObject): void {
+
+        if (!this.isBrowser)
+            return;
+
         try {
             localStorage.setItem(key, JSON.stringify(value));
         } catch (error) {
-            console.error('Error saving to local storage', error);
+            console.error('An error saving to local storage', error);
         }
     }
 
@@ -24,15 +41,17 @@ export class LocalStorageService {
      * @param key 
      * @returns 
      */
-    public getItem(key: string): any {
-        try {
-            const item = localStorage.getItem(key);
-            return item ? JSON.parse(item) : null;
+    public getItem<TObject>(key: string): TObject | undefined {
 
-        } catch (error) {
-            console.error('Error reading an item from local storage', error);
-            return null;
-        }
+        if (!this.isBrowser)
+            return;
+
+        const cachedItem: string | null = localStorage.getItem(key);
+
+        if (!cachedItem)
+            return;
+
+        return Utilities.deserializeFromJson<TObject>(cachedItem);
     }
 
     /**
@@ -40,10 +59,14 @@ export class LocalStorageService {
      * @param key 
      */
     public removeItem(key: string): void {
+
+        if (!this.isBrowser)
+            return;
+
         try {
             localStorage.removeItem(key);
         } catch (error) {
-            console.error('Error removing item from local storage', error);
+            console.error('An error removing item from local storage', error);
         }
     }
 
@@ -51,11 +74,15 @@ export class LocalStorageService {
      * Clear all data in the local storage.
      */
     public clear(): void {
+
+        if (!this.isBrowser)
+            return;
+
         try {
             localStorage.clear();
         }
         catch (error) {
-
+            console.error('An error clearing local storage', error);
         }
     }
 }
